requirements.txt
requests
beautifulsoup4
dnspython
nmap
gvm
configparser
tk
